[Fix][Relax][Torch] Materialize runtime scalar shape values (#20138)
Materialize runtime integer scalars from `torch.export` as Relax
symbolic values and convert them back to rank-zero tensors for fill
operations. Preserve `full`/`full_like` dtype semantics, support direct
`fill`/`fill_` without decomposition, and lower symbolic shape
comparisons as real predicates. Runtime scalar extraction now also
accepts statically one-element integer tensors at higher ranks and
signed/unsigned integer dtypes, preserving negative, unsigned, and
64-bit values. Python scalar fills are materialized in the destination
dtype across `full`, `full_like`, `fill`, `fill_`, `masked_fill`, and
`masked_fill_`, avoiding 32-bit narrowing or precision loss. Symbolic
boolean fills, negation, floor division, and modulo are also supported.

+++ tests/python/relax/test_frontend_dynamo.py
@@ -504,7 +504,7 @@ def test_masked_fill():
         ) -> R.Tensor((256, 256), dtype="float32"):
             with R.dataflow():
                 lv: R.Tensor((256, 256), dtype="float32") = R.full_like(
-                    inp_1, R.const(0, "int32"), dtype=None
+                    inp_1, R.const(0.0, "float32"), dtype=None
                 )
                 lv1: R.Tensor((256, 256), dtype="float32") = R.where(inp_0, lv, inp_1)
                 gv: R.Tensor((256, 256), dtype="float32") = lv1
